Finish ping option
This feature was kind of in there but had no real testing around it.  This
finishes it off, exposes it in the API, and adds integration tests.  In the
process of adding this I started running tests locally with asan.  This helped
me find and fix and bunch of silly lifetime mistakes in the FFI.  This should
get added to CI.