chore(ci): Switch from vagrant to docker for integration tests
Orginally this was written using vagrant VMs in CI for integration tests.
This was done to give control over ICMP on a per host basis.  To run VMs
in GitHub Actions I relied on a little bit of a hack.  They allowed
running VMs in macos runners but no others.  Unsurprisingly they slowly
phased this out.  To get the same behavoir we now lean on iptables to
drop ICMP packets to specific IPs